some bug fixes

+++ controllers/user-controller.js
@@ -4,7 +4,6 @@ const Thought = require('../models/Thought');
 const userController = {
 	// Get all users
 	getAllUsers: async (req, res) => {
-		console.log(req);
 		try {
 			const users = await User.find();
 			res.json(users);
@@ -27,6 +26,7 @@ const userController = {
 	// Create a new user
 	createUser: async (req, res) => {
 		try {
+			console.log(req.body);
 			const user = await User.create(req.body);
 			res.json(user);
 		} catch (error) {
@@ -50,8 +50,15 @@ const userController = {
 	// Delete a user by ID
 	deleteUser: async (req, res) => {
 		try {
-			const user = await User.findByIdAndDelete(req.params.id);
-			res.json(user);
+			const user = await User.findById(req.params.id);
+
+			if (!user) {
+				return res.status(404).json({ message: 'User not found' });
+			}
+			await Thought.deleteMany({ username: user.username });
+			const deletedUser = await User.findByIdAndDelete(req.params.id);
+
+			res.json(deletedUser);
 		} catch (error) {
 			res.status(404).json(error);
 		}
